fix X display requirement for X connection
The silent installer won't work without able to set up an X connection
to a X server in order to load java classes used in the graphic GUI.
Even if we won't use this GUI and do the silent installation from the
command line. So DISPLAY environment variable is added with a working X
server machine.

Also a puppet file resource corresponding to the oracle installation location
file oraInst.loc describing the oracle inventory directory is added.