Corrected include class ordering
Puppet builds a directed graph of resources based on the explicit
relationships you create. There is no guarantee of consistent ordering for resources at the same
level in the graph.

I used a graph inheritance to simulate an ordering on the includes, so
oracle::server::database::enterprise is used for that.