Add stop/start virtual frame buffer
Oracle installer need a xserver even you use the silent install, just
need it to load awt java classes even if they don't use them. So we add
exec resources and file resources to handle the stop and start on
solaris architecture only of the virtual frame buffer.

We also handle to stop processes that use the oracle home directory
before installing the 9.2.0.8 patchset, typically isqlplus, httpd and
tnslnr.

We correct the DISPLAY unix environment variable in the bash profile.