Add exec resource to install 10.2.0.4 software
We add in the class oracle::software exec resource and associated
resources (templates, file resource...) in order to install oracle
database software 10.2.0.1 and the patchset 10.2.0.4 on sparc 64
architecture.

We fix a bug in the wrapper installer script of the patchset 9.2.0.8,
the test on the log file to parse the successful string did not
terminate correctly.

We fix a bug due to the new convention for the oracle major version: it
is now 9.2.0 or 10.2.0

We fix a bug to change the directory path of the software source
location.

We fix a bug on the oracle_product file resource the oracle_dir
definition implied to specify the mode in the file resource.